:recycle: Rename resolver files
