+++ app/src/domain/spot/resolver.ts
