+++ app/src/index.ts
@@ -1,11 +1,12 @@
 import { ApolloServer, AuthenticationError } from "apollo-server";
 import { loadSchemaSync } from "@graphql-tools/load";
 import { GraphQLFileLoader } from "@graphql-tools/graphql-file-loader";
-import { addResolversToSchema } from "@graphql-tools/schema";
+import { addResolversToSchema, chainResolvers } from "@graphql-tools/schema";
 import { join } from "path";
 import admin = require("firebase-admin");
-import { resolvers } from "./resolvers/root";
+import { resolvers } from "./resolvers/resolver";
 import { setUserIDForMe } from "./types/contextHelper";
+import { GraphQLUpload } from "graphql-upload";
 
 admin.initializeApp();
